+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.  
 The format is based on [Keep a Changelog](https://keepachangelog.com/),  
 and this project adheres to [Semantic Versioning](https://semver.org/).
+## [0.10.0] - 2025-01-04
+### Added
+- Preliminary loadsheet
+- Auto Door for Boarding or Deboarding
+
 ## [0.9.0] - 2025-01-04
 ### Added
 - Cargo Door automation